fastapi-web-agent-main1/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.25.2
beautifulsoup4==4.12.3
ollama==0.1.6
fastapi-limiter==0.1.5
redis==4.6.0
groq==0.4.2
python-multipart==0.0.9
aiohttp==3.9.3
lxml==5.1.0
